Adjust poll.py to read multiple (and different) GPIO pins.
It now reads the pins 7, 8, 11, 13, and 15 on the P1 header. It also expects these pins to use pull down resisters with buttons that bridge to 3.3v inputs.

The URL to request on button press is set for each button, though they all point at the same URL until the server they are pinging has proper support for them.